requirements.txt
langchain-text-splitters
pypdf
fitz
fireworks-ai
python-dotenv
google-generativeai
tabulate
sentence-transformers
tenacity